Move get game cover into the component, its only used there
